requirements.txt
Flask==1.1.4
Flask-Bcrypt==0.7.1
Flask-GraphQL==2.0.1
Flask-GraphQL-Auth==1.3.2
Flask-JWT-Extended==3.25.1
Flask-SQLAlchemy==2.5.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
psycopg2-binary==2.9.2
python-dotenv==0.19.2
SQLAlchemy==1.4.27
bcrypt==3.2.0